+++ internal/component/pyroscope/scrape/scrape.go
@@ -274,7 +274,7 @@ func New(o component.Options, args Arguments) (*Component, error) {
 			config_util.WithDialContextFunc(httpData.DialFunc),
 		},
 	}
-	scraper := NewManager(scrapeHttpOptions, alloyAppendable, o.Logger)
+	scraper := NewManager(scrapeHttpOptions, alloyAppendable, o.Logger, o.Registerer)
 	c := &Component{
 		opts:          o,
 		cluster:       clusterData,